libtirpc: Fix warning about SOL_IPV6 redefine
libtirpc: Fix warning about SOL_IPV6 redefine. The question is
whether the original |#define SOL_IPV6 IPPROTO_IPV6| was ever correct.
|IPPROTO_IPV6| is defined as |IPPROTO_IPV6 = 41|,
but |SOL_IPV6| in WinDDK is defined as |SOL_IPV6 = (SOL_SOCKET-5)|,
where |SOL_SOCKET = 0xffff|, |SOL_IPV6 == so 0xffff-5|, which is
clearly not |41|.

But this works:
$ nfs_mount.exe -o sec=sys,port=2049 Z '[fe80::219:99ff:feae:73ce]:/export/home/rmainz'

But... why ?
But in any case the upcoming libtirpc update will just stomp over
the problem, and in the meantime the patch switches the definition
of |SOL_IPV6| to the official DDK value.

Reviewed-by: Cedric Blancher <cedric.blancher@gmail.com>
Signed-off-by: Cedric Blancher <cedric.blancher@gmail.com>
Signed-off-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>