libtirpc: Use fd number instead of SOCKET handle for locking
libtirpc: Use fd number (from |_get_osfhandle()| instead of
|SOCKET| handle for locking.

The original Windows libtirpc implementation used a SOCKET handle and
|WINSOCK_HANDLE_HASH(x) (((x) >> 2) % WINSOCK_HANDLE_HASH_SIZE)| to
get an index to an array of |CONDITION_VARIABLE| to implement locking,
while the POSIX implementation just used the |fd| value as index.
But |SOCKET| can be a 64bit pointer, and the |%| causes wrapound and
re-use of one |CONDITION_VARIABLE| variable for multiple threads -
the symptoms are hangs under hightly parallel loads on multiple
mounts.

The patch now switches the code back to use a |fd| (obtained via
|_open_osfhandle(sock, _O_BINARY)|, |_get_osfhandle(fd)| and our
own |wintirpc_handle2fd(sock)|) as index.

This is also preparation (mostly abstraction) so that we can
update libtirpc to a newer version in the future.

The patch also bumps |FD_SETSIZE| from |128| to |1024|.

Signed-off-by: Cedric Blancher <cedric.blancher@gmail.com>
Signed-off-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>