mount: Fix absolute nfs://-URLs
Fix absolute nfs://-URLs - the '/' after hostport is the separator,
and NOT part of the path, so an absolute path needs another '/'.

Example:
"nfs://hostbar:93/relativepath/a"
"nfs://hostbar:93//absolutepath/a"

nfs://-URLs do support relative paths, usually relative to the
public file handle, but we do not support that yet.

Signed-off-by: Cedric Blancher <cedric.blancher@gmail.com>
Signed-off-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>