deamon: Replace |dprintf()| with |DPRINTF()| which tests debug level inline
Replace the |dprintf()| function with a |DPRINTF()| macro which tests
the debug level inline with the caller.
This saves a whole function call if no debug output is made, plus
parameter setup (which might include additional function calls,
which are only needed for that debug message), and makes debugging or
instrumentation (e.g. DrMemory) faster.

Also, quote strings in debug output (change %s to '%s') and
prefix pointers (change %p to 0x%p) for both |DPRINTF()| and
|eprintf()|.

Signed-off-by: Cedric Blancher <cedric.blancher@gmail.com>
Signed-off-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>