cygwin: nfsd_debug cannot run as user "SYSTEM", fails with exit code 0xC0000022
nfsd_debug.exe cannot run as user "SYSTEM", fails with exit code
0xC0000022 due to file permission errors.

It turns out this is a DLL load error - if a DLL cannot be *read* the
system triggers an exit code of 0xC0000022.

DLLs can only be used if they are readable (chmod a+r), the execute
flag (chmod a+x) only matters for Cygwin, so the bintarball Makefile
target must do a chmod a+rx instead of chmod a+x, as we install the
files usually as a non-SYSTEM plain user.

Reported-by: Martin Wege <martin.l.wege@gmail.com>
Signed-off-by: Cedric Blancher <cedric.blancher@gmail.com>
Signed-off-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>