daemon: Disable 8dot3 ShortName filename generation
Disable 8dot3 ShortName filename generation, the code does currently
not work as expected and |GetShortPathNameW()| in this codepath
eats ~~3% CPU time when building bash.git.

In general Windows is moving away fro generating 8dot3 short names,
but we might need this in the future for compatibility to older
DOS&&Windows applications.

Signed-off-by: Cedric Blancher <cedric.blancher@gmail.com>
Signed-off-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>