cygwin_idmapper.ksh: Handle localised user/group names
Windows uses localised user and group names, e.g. on a
German Windows the group "None" is called "Kein" etc,
which poses a challenge for idmapping.

To fix this we use Cygwin /usr/bin/getent passwd/group
to lookup users&&groups by their SID to get the localised
user/group name.

FIXME: This still means we send a localised group name
to the NFSv4 server, and it needs /etc/group entries for all
localised variations of group "None".
In the future the idmapper should do the mapping in both
directions to avoid this.

Signed-off-by: Cedric Blancher <cedric.blancher@gmail.com>
Signed-off-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>