daemon,tests: Add workaround for Linux nfsd NFSv4.2 SEEK |NFS4ERR_NXIO| bug
Add a workaround for the Linux NFSv4.2 nfsd SEEK bug which returns
|NFS4ERR_NXIO| if it cannot find a data block in a sparse file
(i.e. a file which consists only of a single hole and no data),

The NFSv4.2 RFC says in
https://datatracker.ietf.org/doc/html/rfc7862#section-15.11.3
that "... If the server cannot find a corresponding sa_what,
then the status will still be NFS4_OK, but sr_eof would be
TRUE. ...", but the Linux nfsd SEEK does not behave that way.

Question is... which offset should a conforming NFSv4.2
SEEK_DATA return if there is no data block (i.e. sparse
file which only consists of one hole) ?

Also add tests "hole-only" sparse files and sparse files with
a hole at the end.

Signed-off-by: Cedric Blancher <cedric.blancher@gmail.com>
Signed-off-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>