daemon,sys,tests: RFE: |DeviceIoControl(...,FSCTL_QUERY_ALLOCATED_RANGES,...)| should support |ERROR_MORE_DATA|
RFE: |DeviceIoControl(..., FSCTL_QUERY_ALLOCATED_RANGES, ...)| should
support |ERROR_MORE_DATA| to query the buffer size required to
list all data ranges in |FILE_ALLOCATED_RANGE_BUFFER|s in a
given offset+length.

This is partially ruined by a bug in Windows 10's $ fsutil sparse
queryrange ... # which causes random data to be returned if
|ERROR_MORE_DATA| returns a buffer size different than the buffer
size passed in.

fsutil passes 64 |FILE_ALLOCATED_RANGE_BUFFER| in, and expects
that the size for 64 |FILE_ALLOCATED_RANGE_BUFFER| will be returned
if the error code is |ERROR_MORE_DATA|.
For now we define the cpp symbol
|WINDOWS_FSUTILSPARSEQUERYRANGE_MOREDATA_BUG| to enable our workaround,
until we get a fix for the issue from Microsoft.

Reported-by: Lionel Cons <lionelcons1972@gmail.com>
Signed-off-by: Cedric Blancher <cedric.blancher@gmail.com>
Signed-off-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>