nfs41_build_features.h,sys: Add hack to avoid updating FCB attributes we open an already opened FCB
Add build option (|NFS41_DRIVER_HACK_DISABLE_FCB_ATTR_UPDATE_ON_OPEN|,
"enabled" by default) for a hack to avoid updating FCB attributes if
opened for an already opened FCB.

This is a hack for now, until we can figure out how
to do this correctly (best guess is not to update FCB
attributes if the file is opened for writing, because
the kernel keeps updating the FCB data. The userland
is not affected by this, they get all information from
|nfs41_fcb->BasicInfo| and |nfs41_fcb->StandardInfo|).

We keep this as a build flag for further testing.

The hack passes cthon04, winfstest, building gcc, bash with gcc/clang
and building ms-nfs41-client with Visual Studio 2019.

Without this hack
$ '/cygdrive/c/Program Files/Git/cmd/git' clone ... # will
fail with read errors.

Signed-off-by: Cedric Blancher <cedric.blancher@gmail.com>
Signed-off-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>