daemon: Increase maximum symlink depth to 63
Increase maximum symlink depth from 31 to 63, per
https://learn.microsoft.com/en-us/windows/win32/fileio/reparse-points
---- snip ----
... There is a limit of 63 reparse points on any given path.
NOTE: The limit can be reduced depending on the length of the
reparse point. For example, if your reparse point targets a fully
qualified path, the limit becomes 31.
Windows Server 2003 and Windows XP: There is a limit of 31 reparse
points on any given path.
---- snip ----

Signed-off-by: Cedric Blancher <cedric.blancher@gmail.com>
Signed-off-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>