sys: Remove |MmIsAddressValid()| usage, does not work with Driver verifier
Remove |MmIsAddressValid()| usage, does not work with Driver verifier.

|MmIsAddressValid()| tests whether an adress is "valid" by testing
whether it will cause a page fault.
But when the Driver verifier is active in default mode (e.g.
$ verifier /standard /driver nfs41_driver.sys #) then many more page
faults are generated (cause the verifier tries to track memory
accesses), causing |MmIsAddressValid()| to randomly return |FALSE|.

Root cause is the wrong assuption that |MmIsAddressValid()| will
validate an address, while in reality it reports whether a page fault
will occur for that address.

Signed-off-by: Cedric Blancher <cedric.blancher@gmail.com>
Signed-off-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>