daemon,include,sys: |FileFsAttributeInformation| should always get correct filesystem information
|FileFsAttributeInformation| should always get correct filesystem information
from the userland daemon+NFS server. This is neccesary in case of referrals,
as the referred filesystem might have different properties and capabilities
(i.e. NFSv4.2 filesystem mounted, and the referral is NFSv4.1, or different
maximum name lengths etc.).

On the ToDo list is still a per NFSv4 FSID kernel cache for filesystem
information to avoid updowncalls into the userland daemon if the matching
FSID was already queried.

Signed-off-by: Cedric Blancher <cedric.blancher@gmail.com>
Signed-off-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>