build.vc19: Name cache should handle case-insensitive filesystems
Name cache should handle case-insensitive filesystems.
This implements case-insensitive UTF-8 file name comparisations via
ICU.

Notes:
- This still has issues related to the different Unicode versions
and Unicode implementations on Windows, Windows libicu, NFS server
and the exported filesystem on the NFS server side.
- NTFS is only case-insensitive, but compares details like
Frnech accents 1:1 as is, so we use |ucol_setStrength(cache->icu_coll,
UCOL_SECONDARY)| for the Unicode collator to be compatible with NTFS.
We might want to make this a mount option to choose between
|UCOL_SECONDARY| and |UCOL_PRIMARY|.

Reported-by: Cedric Blancher <cedric.blancher@gmail.com>
Signed-off-by: Cedric Blancher <cedric.blancher@gmail.com>
Signed-off-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>