daemon,include,nfs41_build_features.h,sys: Add hack to override case_preserving+case_insensitive options for testing
Add hack to override case_preserving+case_insensitive options for
testing, which provides two mount options "forcecasepreserving=0/1"
and "forcecaseinsensitive=0/1" to override the
|FATTR4_WORD0_CASE_INSENSITIVE| and/or |FATTR4_WORD0_CASE_PRESERVING|
attributes obtained from the NFS server.

This is only a HACK to circumvent a Linux nfsd bug which always returns
|FATTR4_WORD0_CASE_INSENSITIVE==0|&&|FATTR4_WORD0_CASE_PRESERVING==1|,
even for FAT.

Since Windows file accesses via UNC path make mount options basically
per-server mounts from a single server can only have one set of
"forcecasepreserving=0/1" and "forcecaseinsensitive=0/1" options.

As workaround you can use use the same hostname but a different port
number (as for ms-nfs41-client the port number of part of the UNC path),
e.g. ssh on the NFS server itself to forward port 2050 to 2049 to
pretent it is a different server:
$ ssh -L '*:2050:localhost:2049' root@localhost 'printf "# forwarding...\n" ; sleep $((60*60*24*366*99))' #
and then connect the NFS client to port 2050 on the NFS server.

This build option should be removed as soon as the Linux nfsd bug has
been fixed.

THIS OPTION MUST NOT BE USED ON PRODUCTION SYSTEMS!!

Reported-by: Cedric Blancher <cedric.blancher@gmail.com>
Signed-off-by: Cedric Blancher <cedric.blancher@gmail.com>
Signed-off-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>